Небольшие правки сервисов + Начал писать тесты для сервисов

+++ src/main/java/com/sap/hana/cloud/samples/springboot/service/impl/DefaultCheckPositionService.java
@@ -3,27 +3,36 @@ package com.sap.hana.cloud.samples.springboot.service.impl;
 import com.sap.hana.cloud.samples.springboot.dao.CheckPositionRepository;
 import com.sap.hana.cloud.samples.springboot.model.check.CheckPosition;
 import com.sap.hana.cloud.samples.springboot.service.CheckPositionService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 /**
  * Created by Shishkov A.V. on 05.07.18.
  */
 @Service
 public class DefaultCheckPositionService implements CheckPositionService {
+	@Autowired
 	private CheckPositionRepository repository;
 
-	public DefaultCheckPositionService(CheckPositionRepository repository) {
-		this.repository = repository;
+	@Override
+	public List<CheckPosition> findAll() {
+		return repository.findAll();
 	}
 
 	@Override
 	public CheckPosition findById(Long id) {
-		return repository.findById(id).get();
+		Optional<CheckPosition> result = repository.findById(id);
+		return result.isPresent() ? result.get() : null;
 	}
 
 	@Override
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (findById(id) != null) {
+			repository.deleteById(id);
+		}
 	}
 
 	@Override
